Restructure and rebuild TOS content.
