+++ src/components/shared/About/terms-of-service/TermOfService.module.css
@@ -1,0 +1,64 @@
+.model_container {
+  display: flex;
+  justify-content: center;
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-color: #00000080;
+  z-index: 550;
+}
+
+.model_body {
+  width: 500px;
+  max-width: 500px;
+  height: auto;
+  background-color: #fff;
+  border-radius: 10px;
+  box-shadow: 0 0 10px #0003;
+  padding-top: 20px;
+  padding-bottom: 20px;
+  z-index: 600;
+  cursor: crosshair;
+  overflow-y: auto;
+}
+
+.model_event {
+  width: 100%;
+  height: 100%;
+  position: absolute;
+  background-color: transparent;
+}
+
+.terms_of_service_content {
+  text-align: left;
+  padding: 50px;
+}
+
+.terms_of_service_content h1,
+.terms_of_service_content h2 {
+  color: #1d724b;
+  font-size: 20px;
+}
+
+.terms_of_service_content h2 {
+  border-bottom: 1px solid #ccc;
+  padding-bottom: 5px;
+  margin-bottom: 16px;
+  font-weight: 500;
+}
+
+.terms_of_service_content p {
+  margin-bottom: 20px;
+}
+
+.terms_of_service_content ul {
+  list-style-type: none;
+  margin-bottom: 20px;
+  font-size: 14px;
+}
+
+.terms_of_service_content a {
+  font-weight: bold;
+}
